# Optimize linspace to avoid GPU -> CPU copy.