<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs

### Describe
<!-- Describe what this PR does -->
优化linspace实现，避免输入Start、Stop、Number不必要的GPU <-> CPU之间的传输。
- 优化前nvprof
```
run command: nvprof --profile-from-start off /work/.virtualenvs_cuda10.2/paddle_py38/bin/python /work/benchmark/api/dynamic_tests_v2/linspace.py --task speed --framework paddle --testing_mode dynamic --json_file /work/benchmark/api/tests_v2/configs/linspace.json --config_id 0 --profiler nvprof --backward True --use_gpu True --repeat 1000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   43.15%  5.6490ms      3000  1.8820us  1.8240us  7.8720us  [CUDA memcpy DtoH]
                   18.92%  2.4777ms      2000  1.2380us  1.1840us  7.0080us  void phi::funcs::VectorizedElementwiseKernel<float, phi::CastFuctor<float, float>, int=1, int=1, int=4>(phi::Array<char const * restrict , float>, phi::funcs::VectorizedElementwiseKernel<float, phi::CastFuctor<float, float, float*, phi::CastFuctor<float, float>>, int=1, int=1, int=4>, int, int, float)
                   18.86%  2.4699ms      2000  1.2340us  1.1830us  6.2400us  void phi::funcs::VectorizedElementwiseKernel<float, phi::FullFuctor<float, float>, int=0, int=1, int=4>(phi::Array<char const * restrict , float>, phi::funcs::VectorizedElementwiseKernel<float, phi::FullFuctor<float, float, float*, phi::FullFuctor<float, float>>, int=0, int=1, int=4>, int, int, float)
                    9.61%  1.2581ms      1000  1.2580us  1.1840us  6.1760us  void phi::LinspaceKernelInner<float>(float, phi::LinspaceKernelInner<float>, double, long, phi::LinspaceKernelInner<float>*)
                    9.46%  1.2381ms      1000  1.2380us  1.1840us  7.1360us  void phi::funcs::VectorizedElementwiseKernel<int, phi::FullFuctor<int, int>, int=0, int=1, int=4>(phi::Array<char const * restrict , int>, phi::funcs::VectorizedElementwiseKernel<int, phi::FullFuctor<int, int, int*, phi::FullFuctor<int, int>>, int=0, int=1, int=4>, int, int, int)

total gpu_time: 13.0915 ms
```

- 优化后nvprof
```
run command: nvprof --profile-from-start off /work/.virtualenvs_cuda10.2/paddle_py38/bin/python /work/benchmark/api/dynamic_tests_v2/linspace.py --task speed --framework paddle --testing_mode dynamic --json_file /work/benchmark/api/tests_v2/configs/linspace.json --config_id 0 --profiler nvprof --backward True --use_gpu True --repeat 1000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:  100.00%  1.2648ms      1000  1.2640us  1.2150us  7.3600us  void phi::LinspaceKernelInner<float>(float, phi::LinspaceKernelInner<float>, double, long, phi::LinspaceKernelInner<float>*)

total gpu_time: 1.2648 ms
```